Don't split codefix and analyzer tests
